apps/product/admin.py
from django.contrib import admin
from .models import Product, Category
from mptt.admin import DraggableMPTTAdmin


@admin.register(Category)
class CategoryAdmin(DraggableMPTTAdmin):
    list_display = ('tree_actions', 'indented_title',)
    prepopulated_fields = {'slug': ('title',)}
    list_display_links = ('indented_title',)


@admin.register(Product)
class ProductAdmin(admin.ModelAdmin):
    prepopulated_fields = {'slug': ('title',)}
